3i/requirements.txt
# Core utilities
numpy
scipy
sounddevice
rich

# Hardware control
RPi.GPIO
Adafruit-SSD1306
Pillow
picamera2

# Audio + speech
vosk
openai-whisper
piper-tts

# Vision + detection
ultralytics

# LLM + RAG
llama-cpp-python
chromadb

